% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.decay.stats}
\alias{calc.decay.stats}
\title{Calculate model summary and validation statistics for trip duration decay model}
\usage{
calc.decay.stats(x, y.route, districts, pop.dens, dist.mat)
}
\arguments{
\item{x}{JAGS model object}

\item{y.route}{matrix of route-level trip duration counts}

\item{districts}{vector of district names}

\item{pop.dens}{vector of district population densites}

\item{dist.mat}{distance matrix}
}
\value{
a list containing a dataframe of parameter estimates and diagnostics, and a matrix of actual and predicted trip duration counts
}
\description{
This function takes the output from a the \code{jags.model} function and calculates summaries of
the posterior parameter estimates, convergence diagnostics, and sample sizes for each route.
}
\examples{
load('./data/duration_data_arrays_5day_full.Rdata')                    # y.month, y.route, y.pop
load('./MARCC/output/decay_model_route_hierarchical_5day_full.Rdata')  # out # Decay model parameters (Lambda)
load('./data/distance_matrix_named.rdata')                             # distance.matrix
pop <- read.csv("./data/population_data.csv")                          # population densiity

districts <- dimnames(y.route)$origin

dens <- vector()
for (i in seq_along(districts)) {
     dens <- c(dens, pop$density[as.numeric(which(pop$ID_2 == districts[i]))])
}

mod.decay <- calc.decay.stats(x=out, 
                              y.route=y.route, 
                              districts=districts, 
                              pop.dens=dens, 
                              dist.mat=distance.matrix[districts, districts])
}
\seealso{
Other model processing: \code{\link{get.param.vals}()},
  \code{\link{mob.data.array.pop.level}()},
  \code{\link{mob.data.array.route.level}()}
}
\author{
John Giles
}
\concept{model processing}
