% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.hpd}
\alias{calc.hpd}
\title{Calculate HPD of aggregated waiting time distributions}
\usage{
calc.hpd(x)
}
\arguments{
\item{x}{aggregated waiting time distributions for each district (output from \code{\link{calc.wait.time}})}
}
\value{
a dataframe
}
\description{
This function calculates the highest posterior density (HPD) interval for the aggregated 
waiting time distributions returned by the \code{\link{calc.wait.time}} function. The function 
calculates the maximum of each aggregated probability distribution along with its 50 and 95 percent HPD.
}
\examples{
load("./output/TSIR_sim_flu_high.Rdata")

x <- sim$B$wait.time   # basic gravity model
#x <- sim$R$wait.time   # gravity model with duration 

agg <- agg.wait.time(x)
hpd <- calc.hpd(agg)

par(mfrow=c(3,4))
for (i in 1:nrow(hpd)) {
     plot(agg1[i,], type='l', main=paste("district ", hpd$district[i]), ylab='Waiting time density')
     abline(v=hpd$lo95[i], lty=2, col='red')
     abline(v=hpd$hi95[i], lty=2, col='red')
     abline(v=hpd$max[i], col='red')
}

}
\seealso{
Other simulation: \code{\link{calc.p}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
