% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.prop.imm}
\alias{calc.prop.imm}
\title{Calculate the proportion of population immune}
\usage{
calc.prop.imm(v1, v2)
}
\arguments{
\item{v1}{a scalar giving the proportion population vaccinated with first dose}

\item{v2}{a scalar giving the proportion population vaccinated with second dose}
}
\value{
A dataframe containing the relative proportions of the population that have recieved 0, 1, or 2 doses
}
\description{
This function calculates the proportion immune using conditional probabilities. The method 
assumes that vaccination events are dependent, where individuals that have recieved the first
dose recieve the second dose before those that have missed the first dose.
}
\examples{
calc.prop.imm(0.9, 0.8) # Those who get first dose most likely to get second

calc.prop.imm(1,1) # Everyone gets both doses (yay!)

props <- calc.prop.imm(rbeta(1,3,1), rbeta(1,2,2)) 
sum(props[1:2,2]) # total proportion immune







}
\seealso{
Other susceptibility: \code{\link{get.age.beta}},
  \code{\link{get.beta.params}}
}
\author{
John Giles
}
\concept{susceptibility}
