% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.prop.inf}
\alias{calc.prop.inf}
\title{Calculate proportion infected in each district}
\usage{
calc.prop.inf(x, N)
}
\arguments{
\item{x}{matrix of total infected for all districts in simulation}

\item{N}{vector of population size of each district}
}
\value{
a dataframe
}
\description{
This function takes the output from a TSIR simulation and calculates the mean 
proportion infected in each district.
}
\examples{
load("./output/TSIR_sim_influenza_high.Rdata") # sim
load('./data/N_pop.rdata')                     # N.pop

# Get only relevant districts
districts <- colnames(sim$B$tot.inf)
N <- N[which(N$ID \%in\% districts), 'N']

# Calculate mean proportion infected in each district
prop.inf.basic <- calc.prop.inf(sim$B$tot.inf, N=N)
prop.inf.duration <- calc.prop.inf(sim$R$tot.inf, N=N)
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}}, \code{\link{load.obj}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
