% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.prop.remain}
\alias{calc.prop.remain}
\title{Proportion of individuals remaining for full epidemic generation}
\usage{
calc.prop.remain(d, gen.t, sub.samp = NULL)
}
\arguments{
\item{d}{a three- or four-dimensional array containing route- or month-level trip duration counts produced by the \code{\link{mob.data.array}} function. 
Note that \code{calc.prop.remain} assumes that the duration data array is NOT aggregated (e.g. \code{mob.data.array} argument \code{agg.int}=1)}

\item{gen.t}{the time interval in days used to define the epidemic generation}

\item{sub.samp}{scalar indicating the number of generations to subsample, if NULL (default), will use all observed generation times in the data (which will increase computation time for large numbers of locations)}
}
\value{
if \code{d} is a month-level duration data array, then a 4D array with values between 0 and 1 is returned. If \code{d} is a route-level duration data array, 
then returns a 3D array is returned.
}
\description{
This function calculates the proportion of individuals that remain in a location for
all of epidemic generation \eqn{n}. The value is represented by the variable \eqn{p_jt}, which
is defined as the effective probability if individuals that stay for a full epidemic generation 
when they travel to destination \eqn{j} at time \eqn{t}:\cr
\cr
\eqn{p_ij} = Pr(remaining for all of \eqn{n^th} epidemic generation | generation time \eqn{g})\cr
\cr
Because the Namibia mobility data spans 4 years there are many potential time intervals for which to calculate 
the proportion of individuals remaining for full epidemic generation. The \code{sub.samp} argument randomly 
selects X number of these generations to reduce computation time.
}
\examples{

# Duration data for the purpose of subsetting districts
load('./data/duration_data_arrays_1day_full.Rdata') 

# Proportion travellers remaining for full epidemic generation based on subsample of 10, where the generation time is 14 days (measles)
p1 <- calc.prop.remain(d=y.route, gen.t=14, sub.samp=10)
p2 <- calc.prop.remain(d=y.month, gen.t=14, sub.samp=10)

# Get destination-level mean and variance
p1.mean <- apply(p1, 2, mean, na.rm=TRUE)
p1.var <- apply(p1, 2, var, na.rm=TRUE)

p2.mean <- apply(p2, 2, mean, na.rm=TRUE)
p2.var <- apply(p2, 2, var, na.rm=TRUE)
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}}, \code{\link{load.obj}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
