% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.prop.route.type}
\alias{calc.prop.route.type}
\title{Calculate observed proportion of each route type}
\usage{
calc.prop.route.type(d, hi, lo)
}
\arguments{
\item{d}{A longform data.frame produced by the _ function, which contains trip duration data}

\item{hi}{A vector of numerical district IDs in the high population density group}

\item{lo}{A vector of numerical district IDs in the low population density group}
}
\value{
A dataframe with four rows with values between 0 and 1, rows sum to 1
}
\description{
This function calculates the observed proportion of each of the four route types for 
each day in the trip duration data. The route types are defined by the population density 
of the origin and destination districts:
\enumerate{
  \item High density to high density (HH)
  \item High density to low density (HL)
  \item Low density to high density (LH)
  \item Low density to low density (LL)
}
}
\examples{
d <- read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform_metadata_12_dists_hi_lo_dens.csv", stringsAsFactors=F)

test <- calc.prop.route.type(d=d, 
                             hi=c(42,13,3,87,34), 
                             lo=c(10,15,14,48,29,103,64))

boxplot.matrix(as.matrix(test[,2:5]), col='lightblue')

}
\seealso{
Other data synthesis: \code{\link{calc.prop.tot.trips}},
  \code{\link{calc.samp.size}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.district.pop}},
  \code{\link{get.duration.counts}},
  \code{\link{get.holidays}}, \code{\link{get.subsamp}},
  \code{\link{get.xy.counts}},
  \code{\link{mob.data.array}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
