% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.prop.vacc.SIA}
\alias{calc.prop.vacc.SIA}
\title{Calculate vaccination coverage given routine vaccination and a supplementary campaign}
\usage{
calc.prop.vacc.SIA(v1, v2, v3 = NULL, S)
}
\arguments{
\item{v1}{a scalar giving the proportion vaccinated with first routine immunization}

\item{v2}{a scalar giving the proportion vaccinated with second routine immunization}

\item{v3}{a scalar giving the proportion vaccinated with third routine immunization (default = NULL)}

\item{S}{a scalar giving the proportion vaccinated with supplemental campaign}
}
\value{
A dataframe containing the relative proportions of the population that have recieved 0, 1, 2, 3, or 4 doses
}
\description{
This function calculates the proportion immune by calculating the conditional probability of routine
vaccination with two- or three-dose routine coverage and one SIA campaign. The method 
assumes that vaccination events are dependent, where individuals that have recieved the first
dose are the most likely to recieve the second dose and those that have received both the first 
and second doses are the most likely to receive the third. Receipt of a dose through the SIA campaign
is dependent upon vaccination with at least one dose prior to the SIA campaign.
}
\details{
When \code{v3 = NULL}, the function uses the simpler two dose method.
}
\examples{
calc.prop.vacc.SIA(v1=0.9, v2=0.8, S=0.97) # High coverage two-dose routine and campaign

calc.prop.vacc.SIA(v1=0.9, v2=0.85, v3=0.8, S=0.95) # High coverage two-dose routine and campaign

calc.prop.vacc.SIA(v1=0.5, v2=0.5, v3=0.5, S=0.95) # Low routine coverage and high campaign

# Boundary conditions
calc.prop.vacc.SIA(v1=1, v2=1, v3=1, S=1)
calc.prop.vacc.SIA(v1=1, v2=0, v3=0, S=0)

# Complex models reduce to simpler models
calc.prop.vacc.SIA(v1=0.9, v2=0.8, v3=0.7, S=0)
calc.prop.vacc(v1=0.9, v2=0.8, v3=0.7)

calc.prop.vacc.SIA(v1=0.9, v2=0, v3=0, S=0)
calc.prop.vacc(v1=0.9, v2=0, v3=0)
calc.prop.vacc(v1=0.9, v2=0)
}
\seealso{
Other susceptibility: 
\code{\link{calc.prop.vacc}()},
\code{\link{get.age.beta}()},
\code{\link{get.beta.params}()}
}
\author{
John Giles
}
\concept{susceptibility}
