% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{calc.route.type}
\alias{calc.route.type}
\title{Calculate observed number of trips of each route type}
\usage{
calc.route.type(m, hi, lo, per.route = FALSE)
}
\arguments{
\item{m}{A 3-dimensional data array produced by the \code{\link{mob.data.array}} function containing total trip counts (dimensions are origin, destination, time)}

\item{hi}{A vector of numerical district IDs in the high population density group}

\item{lo}{A vector of numerical district IDs in the low population density group}

\item{per.route}{logical indicating whether or not to scale number of trips by the number of routes in each route type}
}
\value{
A five column dataframe. If \code{per.route = FALSE}, then counts represent raw number of total trips.  If \code{per.route = TRUE}, then counts represent total trips per route.
}
\description{
This function calculates the observed number of trips of each of the four route types for 
each temporal unit in a mobility data array. The route types are defined by the population density 
of the origin and destination districts:
\enumerate{
  \item High density to high density (HH)
  \item High density to low density (HL)
  \item Low density to high density (LH)
  \item Low density to low density (LL)
}
}
\examples{
# Subset of 5 high pop density and 7 low pop density
d <- read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform_metadata_12_dists_hi_lo_dens.csv", stringsAsFactors=F)
pop <- read.csv("./data/population_data.csv")

cutoff <- 980
hi <- pop$ID_2[pop$density >= cutoff]
lo <- pop$ID_2[pop$density < cutoff]

m <- mob.data.array(orig=d$from,
                    dest=d$to,
                    time=d$day, # total trip counts for each unique day
                    count=d$count,
                    name='movement')
str(m)

par(mfrow=c(1,2))

p1 <- calc.route.type(m, hi, lo, per.route=FALSE) # Raw number of trips in each route type

plotrix::violin_plot(p1[,2:5], violin_width = 1, box_width=0.05, 
                     xlab='Route type', ylab='Raw total number trips', main=NULL,
                     x_axis_labels=c('High to high', 'High to low', 'Low to high', 'Low to low'))

p2 <- calc.route.type(m, hi, lo, per.route=TRUE) # Total number of trips per route

plotrix::violin_plot(p2[,2:5], violin_width = 1, box_width=0.05, 
                     xlab='Route type', ylab='Total trips per route', main=NULL,
                     x_axis_labels=c('High to high', 'High to low', 'Low to high', 'Low to low'))






}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.sparse.mob.matrix}()},
\code{\link{get.stay.data}()},
\code{\link{get.subsamp}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
