% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.samp.size}
\alias{calc.samp.size}
\title{Get the sample size of a route-level \code{mob.data.array} matrix object}
\usage{
calc.samp.size(x)
}
\arguments{
\item{x}{a three dimensional array produced by the \code{mob.data.array.route.level} function}
}
\value{
a matrix of sample sizes for each \eqn{ij} route
}
\description{
This function takes the output from a the \code{mob.data.array.route.level} function and calculates
the sample size of observations for each \eqn{ij} route.
}
\examples{
load('./data/duration_data_arrays_3day_full.Rdata') # y.month, y.route, y.pop 

samp.size <- calc.samp.size(y.route)
}
\seealso{
Other data synthesis: \code{\link{calc.prop.route.type}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.district.pop}},
  \code{\link{get.duration.counts}},
  \code{\link{get.holidays}}, \code{\link{get.subsamp}},
  \code{\link{get.xy.counts}},
  \code{\link{mob.data.array}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
