% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.wait.time}
\alias{calc.wait.time}
\title{Calculate aggregate waiting time distribution}
\usage{
calc.wait.time(x)
}
\arguments{
\item{x}{a three-dimensional array containing waiting time distributions for all districts and simulations}
}
\value{
a matrix (rows = districts, cols = epidemic generation)
}
\description{
This function aggregates the all simulated waiting time distributions into a single probability density for each district. 
The method uses a simple linear combination of waiting time probabilities, sometimes referred to as the 'linear opinion pool'.
}
\examples{
load("./output/TSIR_sim_flu_high.Rdata")

x1 <- sim$B$wait.time
x2 <- sim$R$wait.time

agg1 <- calc.wait.time(x1)
agg2 <- calc.wait.time(x2)

plot(agg1[1,], type='l', col='blue', cex=2)
lines(agg2[1,], col='red', cex=2)
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}}, \code{\link{calc.p}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{decay.func}}, \code{\link{get.beta.params}},
  \code{\link{load.obj}}, \code{\link{sim.TSIR.full}},
  \code{\link{sim.TSIR}}, \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
