% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{check.distance.matrix}
\alias{check.distance.matrix}
\title{Fix off-diagonal zeros in distance matrix}
\usage{
check.distance.matrix(D)
}
\arguments{
\item{D}{distance matrix produced by the \code{\link{get.distance.matrix}} function}
}
\value{
the same distance matrix with zeros fixed
}
\description{
This function checks the distance matrix produced by the \code{\link{get.distance.matrix}} function for off-diagonal zeros. The function
fills in zeros with nearby locations from the same admin unit.
}
\seealso{
Other utility: 
\code{\link{get.admin.level}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.coords}()},
\code{\link{get.unique.names}()},
\code{\link{load.obj}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
