% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{decay.func}
\alias{decay.func}
\title{Exponential decay function}
\usage{
decay.func(N0, lambda, y)
}
\arguments{
\item{N0}{intercept (baseline number of expected trips at \eqn{y=0})}

\item{lambda}{decay rate}

\item{y}{a scalar or vector giving the data value(s) (e.g. trip duration)}
}
\value{
A scalar or vector of \eqn{N(y)}
}
\description{
This function calculates exponential decay of a data value \code{y} (e.g. trip duration)
given the intercept \eqn{N0} and decay rate \eqn{\lambda} parameters using the functional form:
\eqn{N(y) = N0 exp(-\lambda*y)}. If \eqn{y =} duration, the function calculates the number of trips counted for a given duration.
}
\seealso{
Other simulation: 
\code{\link{calc.hpd}()},
\code{\link{calc.prop.inf}()},
\code{\link{calc.prop.remain}()},
\code{\link{calc.timing.magnitude}()},
\code{\link{calc.wait.time}()},
\code{\link{get.age.beta}()},
\code{\link{get.beta.params}()},
\code{\link{sim.TSIR.full}()},
\code{\link{sim.TSIR}()},
\code{\link{sim.combine.dual}()},
\code{\link{sim.combine}()},
\code{\link{sim.gravity.duration}()},
\code{\link{sim.gravity}()},
\code{\link{sim.lambda}()},
\code{\link{sim.pi}()},
\code{\link{sim.rho}()},
\code{\link{sim.tau}()}
}
\author{
John Giles
}
\concept{simulation}
