% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{exp.decay}
\alias{exp.decay}
\title{Exponential decay function}
\usage{
\method{exp}{decay}(beta, lambda, y)
}
\arguments{
\item{beta}{intercept (baseline number of expected trips at \eqn{y=0})}

\item{lambda}{decay rate}

\item{y}{a scalar or vector giving the data value(s) (e.g. trip duration)}
}
\value{
A scalar or vector of \eqn{N(y)}
}
\description{
This function calculates exponential decay of a data value \code{y} (in this case duration)
given the intercept \eqn{\beta} and decay rate \eqn{\lambda} parameters using the functional form:
\eqn{N(y) = \beta exp(-\lambda*y)}. If \eqn{y =} duration, the function calculates the number of trips counted for a given duration.
}
\seealso{
Other simulation: \code{\link{calc.p}},
  \code{\link{get.beta.params}}
}
\author{
John Giles
}
\concept{simulation}
