% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_model_funcs.R
\name{fit.jags}
\alias{fit.jags}
\title{Fit model using JAGS}
\usage{
fit.jags(
  jags.data,
  jags.model,
  params,
  n.chain = 2,
  n.burn = 1000,
  n.samp = 1000,
  n.thin = 1,
  DIC = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{jags.data}{a list of data objects}

\item{jags.model}{character string giving the model specification in JAGS/BUGS syntax}

\item{params}{vector of parameters to monitor}

\item{n.chain}{number of MCMC sampling chains}

\item{n.burn}{number of iterations to discard before sampling of chains begins (burn in)}

\item{n.samp}{number of iterations to sample each chain}

\item{n.thin}{interval to thin samples}

\item{DIC}{logical indicating whether or not to calculate the Deviance Information Criterion (DIC)}

\item{parallel}{logical indicating whether or not to run MCMC chains in parallel or sequentially (default = FALSE)}
}
\value{
an \code{\link[coda:mcmc.list]{mcmc.list}} object
}
\description{
This function fits
}
\seealso{
Other model: 
\code{\link{fit.gravity}()},
\code{\link{fit.prob.travel}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{summarize.hmob}()}
}
\author{
John Giles
}
\concept{model}
