% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.admin.level}
\alias{get.admin.level}
\title{Find the lowest admin unit}
\usage{
get.admin.level(data)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel.survey}} or derivative thereof}
}
\value{
integer
}
\description{
This function checks which admin levels are in a generalized a data frame formatting like \code{\link{travel.survey}} and 
returns the lowest admin unit.
}
\seealso{
Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.coords}()},
\code{\link{get.unique.names}()},
\code{\link{load.obj}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
