% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.age.beta}
\alias{get.age.beta}
\title{Get Beta parameters for age groups}
\usage{
get.age.beta(age = NULL, vacc = NULL, breaks = NULL)
}
\arguments{
\item{age}{a vector giving the age at vaccination of each individual}

\item{vacc}{a binary vector indicating the vaccination status of each individual}

\item{breaks}{a scalar or vector of age group breaks (passed to \code{\link{.bincode}}). If NULL (default), calculates Beta parameters for all data together.}
}
\value{
A dataframe containing the sample size, mu, sigma, and Beta distribution paramters for each age group
}
\description{
This function finds the two shape parameters for the Beta distribution for aggregated age groups given individual-level 
vaccination data. The individual-level data is comprised of the age of the individual (\code{age}) and a binary indicator of vaccination status 
(\code{vacc}). The function first uses a Binomial GAM to estimate the mean (mu) and variance (sigma) of the proportion vaccinated for each 
of the defined age groups and then finds the shape parameters of the Beta distribution analytically using \code{\link{get.beta.params}}.
}
\details{
Note that a Binomial GLM is used if the number of age groups is 2 or less.
}
\examples{
# DHS vaccination coverage file downloaded from the website 
dhs <- read.csv('./data/Namibia2013.csv', stringsAsFactors=FALSE) 

# For all observations together (population mean)
tmp <- get.age.beta(age=dhs$age.in.months,
                    vacc=dhs$measles.y,
                    breaks=NULL)


# Two age groups < 12 months and > 12 months
tmp <- get.age.beta(age=dhs$age.in.months,
                    vacc=dhs$measles.y,
                    breaks=12)

par(mfrow=c(1,2))
for(i in 1:2) curve(dbeta(x, tmp$shape1[i], tmp$shape2[i]), 0, 1)


# 6-month age groups
tmp <- get.age.beta(age=dhs$age.in.months,
                    vacc=dhs$measles.y,
                    breaks=seq(0, 60, 6))

par(mfrow=c(2,5))
for(i in 1:10) curve(dbeta(x, tmp$shape1[i], tmp$shape2[i]), 0, 1)


# Each unique age in months
tmp <- get.age.beta(age=dhs$age.in.months,
                    vacc=dhs$measles.y,
                    breaks=seq(0, 60, 1))

par(mfrow=c(1,1))
plot(tmp$mu, type='l')


# 3-month age groups for each region
tmp <- foreach(i=unique(dhs$region.residence), .combine='rbind') \%do\% {
     
     sel <- dhs$region.residence == i
     
     cbind(
          data.frame(region=i),
          get.age.beta(age=dhs$age.in.months[sel],
                       vacc=dhs$measles.y[sel],
                       breaks=seq(0, 60, 3))
     )
}
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}

Other susceptibility: \code{\link{calc.prop.vacc}},
  \code{\link{get.beta.params}}
}
\author{
John Giles
}
\concept{simulation}
\concept{susceptibility}
