% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.beta.params}
\alias{get.beta.params}
\title{Get parameters for Beta distribution}
\usage{
get.beta.params(mu = NULL, sigma = NULL, quantiles = NULL,
  probs = NULL)
}
\arguments{
\item{mu}{scalar giving the mean \eqn{\mu}}

\item{sigma}{scalar giving the variance \eqn{\sigma^2}}

\item{quantiles}{vector of quantiles for which proportions are observed}

\item{probs}{vector of observed proportions}
}
\value{
A list containing the two shape parameters of the Beta distribution
}
\description{
This function to calculates the two shape parameters for the Beta distribution 
using EITHER the mean \eqn{\mu} and variance \eqn{\sigma^2} of a random variable between 0 and 1. 
When the mean and variance are given, the solution is found analytically. When the proportions 
(or probabilities) at each quantile are given, the solution is found numerically using the Nelder-Mead 
optimiation algorithm. Note that the function expects only \code{mu} and \code{sigma} OR \code{quantiles} 
and \code{probs} as arguments.
}
\examples{

# Duration data for the purpose of subsetting districts
load('./data/duration_data_arrays_1day_full.Rdata') 

# Proportion travellers remaining for full epidemic generation based on subsample of 10, where the generation time is 14 days (measles)
p1 <- calc.prop.remain(d=y.route, gen.t=14, sub.samp=10)
p2 <- calc.prop.remain(d=y.month, gen.t=14, sub.samp=10)

# Get destination-level mean and variance
p1.mean <- apply(p1, 2, mean, na.rm=TRUE)
p1.var <- apply(p1, 2, var, na.rm=TRUE)

p2.mean <- apply(p2, 2, mean, na.rm=TRUE)
p2.var <- apply(p2, 2, var, na.rm=TRUE)
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{load.obj}}, \code{\link{sim.TSIR.full}},
  \code{\link{sim.TSIR}}, \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
