% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.crossdist}
\alias{get.crossdist}
\title{Get distance between two points for one observation}
\usage{
get.crossdist(xy1, xy2, id1 = NULL, id2 = NULL)
}
\arguments{
\item{xy1}{two column matrix of XY coordinates for first group}

\item{xy2}{two column matrix of XY coordinates for second group}

\item{id1}{optional names for first group}

\item{id2}{optional names for second group}
}
\value{
numeric scalar or matrix
}
\description{
Takes X and Y coordinates of two locations and returns cross distance for all entries.
}
\examples{
# Some XY coords in decimal degrees
xy1 <- data.frame(x=rnorm(3, 100, 1), 
                 y=rnorm(3, 20, 1),
                 id=LETTERS[1:3])

xy2 <- data.frame(x=rnorm(5, 100, 1), 
                 y=rnorm(5, 20, 1),
                 id=LETTERS[4:8])

D <- get.crossdist(xy1=xy1[,1:2],
                   xy2=xy2[,1:2],
                   id1=xy1[,3],
                   id2=xy2[,3])

D
D*111.35 # in km


# Overlapping sets of coordinates
xy1 <- data.frame(x=rnorm(3, 100, 1), 
                  y=rnorm(3, 20, 1),
                  id=LETTERS[1:3])

xy2 <- rbind(xy1,
             data.frame(x=rnorm(5, 100, 1), 
                        y=rnorm(5, 20, 1),
                        id=LETTERS[4:8]))

D <- get.crossdist(xy1=xy1[,1:2],
                   xy2=xy2[,1:2],
                   id1=xy1[,3],
                   id2=xy2[,3])

D
D*111.35 # in km
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.sparse.mob.matrix}()},
\code{\link{get.subsamp}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
