% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.distance.class}
\alias{get.distance.class}
\title{Add distance between origin and destination districts to long form data set}
\usage{
get.distance.class(a, b)
}
\arguments{
\item{a}{expects output from \code{parse.long.form} (trip_durations_longform.csv)}

\item{b}{expects data object \code{d.meta.3} (route_classification.csv)}
}
\value{
dataframe
}
\description{
This function takes output from the \code{parse.long.form} function and appends columns containing distance between origin and destination districts.
This function also adds two columns indicating the commuting classification of the travel:
\enumerate{
     \item \code{rcl_same_adm1} = whether trip was within same admin unit
     \item \code{cl_contig} = whether trip was within same, to adjacent, or to non-adjacent admin unit
     }
}
\examples{
# Add district population sizes and route classifications 
tmp <- get.distance.class(
     a=read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform.csv"),
     b=read.csv("BeyondCommuting2/data/route_classification.csv")
)
head(tmp)
}
\seealso{
Other data synthesis: \code{\link{calc.prop.tot.trips}()},
  \code{\link{calc.route.type}()},
  \code{\link{calc.samp.size}()},
  \code{\link{get.distance.counts}()},
  \code{\link{get.district.names.xy}()},
  \code{\link{get.district.pop}()},
  \code{\link{get.duration.counts}()},
  \code{\link{get.holidays}()},
  \code{\link{get.subsamp}()},
  \code{\link{get.xy.counts}()},
  \code{\link{mob.data.array}()},
  \code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
