% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.distance.counts}
\alias{get.distance.counts}
\title{Get matrix of trip distance counts}
\usage{
get.distance.counts(
  a,
  t,
  d,
  type = "matrix",
  distID = districtIDs,
  n.cores = NULL
)
}
\arguments{
\item{a}{origin district (can take integer ID or character name)}

\item{t}{character string indicating the temporal interval the count trip distances (e.g. 'month', 'week', 'doy')}

\item{d}{longform mobility data, expects \code{d.42} data object (trip_durations_longform_metadata_42.csv) or similar}

\item{type}{return a 'matrix' or a 'dataframe' with counts in longform, default = 'matrix'}

\item{distID}{district IDs and names, expects \code{districtIDs} data object (NamNames.csv), default = districtIDs}

\item{n.cores}{number cores to use when parallel computing (default = NULL, which uses half available cores)}
}
\value{
matrix or dataframe
}
\description{
The \code{get.distance.counts} function calculates a matrix or long-form dataframe of trip distance counts for a temporal 
interval \code{t} given an origin \code{a}. Only takes origin district because it uses distances across all destinations.
}
\examples{
# Distribution of trip distances that originate from Windhoek East (42)

data(d.42) # same as ./BeyondCommuting2/trip_lengths/trip_durations_longform_metadata_42.csv
data(districtIDs) # same as NamNames.csv
orig <- "Windhoek East"

m <- get.distance.counts(a=orig, t='month', d=d.42, n.cores=4)

see(m)

lattice::wireframe(m[,1:50],
                   xlab = "Month", ylab = "Distance", zlab='Freq',
                   drape = T, colorkey = F,
                   light.source = c(10,0,10), 
                   col.regions = colorRampPalette(c("blue", "red"))(100),
                   screen = list(z = 230, x = -60))
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.sparse.mob.matrix}()},
\code{\link{get.stay.data}()},
\code{\link{get.subsamp}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
