% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.distance.matrix}
\alias{get.distance.matrix}
\title{Build distance matrix from XY coordinates}
\usage{
get.distance.matrix(x, y, id)
}
\arguments{
\item{x}{vector giving X coordinates}

\item{y}{vector giving Y coordinates}

\item{id}{vector of names for each location}
}
\value{
a named matrix of pairwise distances among locations
}
\description{
This function builds the pairwise distance matrix from vectors of XY coordinates and associated names.
}
\examples{
# Some XY coords in decimal degrees
xy <- data.frame(x=rnorm(10, 100, 1), 
                 y=rnorm(10, 20, 1),
                 id=LETTERS[1:10])

D <- get.distance.matrix(x=xy[,1],
                         y=xy[,2],
                         id=xy[,3])

see(D)
see(D*111.35) # in km
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.sparse.mob.matrix}()},
\code{\link{get.subsamp}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
