% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.district.names.xy}
\alias{get.district.names.xy}
\title{Add district names and coordinates to long form data set}
\usage{
get.district.names.xy(a, b)
}
\arguments{
\item{a}{expects output from \code{parse.long.form} (trip_durations_longform.csv)}

\item{b}{expects data object \code{d.meta.1} (namibia_adm2_centroids_name_number_mapping.csv)}
}
\value{
dataframe
}
\description{
This function takes output from the \code{parse.long.form} function and appends columns containing district names and coordinates
}
\examples{
# Add district names and coordinates
t <- Sys.time()
tmp <- get.district.names.xy(
     a=read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform.csv"), 
     b=read.csv("BeyondCommuting2/data/namibia_adm2_centroids_name_number_mapping.csv")
)
head(tmp)
}
\seealso{
Other data synthesis: \code{\link{calc.prop.tot.trips}},
  \code{\link{calc.route.type}},
  \code{\link{calc.samp.size}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.pop}},
  \code{\link{get.duration.counts}},
  \code{\link{get.holidays}}, \code{\link{get.subsamp}},
  \code{\link{get.xy.counts}},
  \code{\link{mob.data.array}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
