% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.district.pop}
\alias{get.district.pop}
\title{Add district population sizes to long form data set}
\usage{
get.district.pop(a, b)
}
\arguments{
\item{a}{expects output from \code{parse.long.form} (trip_durations_longform.csv)}

\item{b}{expects data object \code{d.meta.2} (namibia_pop_from_worldpop_adm2.csv)}
}
\value{
dataframe
}
\description{
This function takes output from the \code{parse.long.form} function and appends columns containing population size
}
\examples{
# Add district population sizes
tmp <- get.district.pop(
     a=read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform.csv"), 
     b=read.csv("BeyondCommuting2/data/namibia_pop_from_worldpop_adm2.csv")
)
head(tmp)
}
\seealso{
Other data synthesis: \code{\link{calc.prop.route.type}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.duration.counts}},
  \code{\link{get.holidays}}, \code{\link{get.xy.counts}},
  \code{\link{mob.data.array}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
