% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.gamma.params}
\alias{get.gamma.params}
\title{Get parameters of Gamma distribution}
\usage{
get.gamma.params(mu, sigma)
}
\arguments{
\item{mu}{the desired mean of the Gamma distribution}

\item{sigma}{the desired standard deviation of the Gamma distribution}
}
\value{
a named numeric vector giving the \code{shape} and \code{rate} parameters of the Gamma distribution
}
\description{
A function that finds the \code{shape} and \code{rate} parameters required by the Gamma distribution given the observed 
mean \code{mu} and standard deviation \code{sigma} of the response variable. Parameters are found numerically using a 
two-dimensional Nelder-Mead optimization algorithm.
}
\examples{
prm <- get.gamma.params(mu=15.3, sigma=9.3)

prm[1]/prm[2] # mean
sqrt(prm[1]/(prm[2]^2)) # sd

plot(density(rgamma(n=1000, shape=prm[1], rate=prm[2])), main='')
abline(v=15.3, lty=2, col='red')

prm <- get.gamma.params(mu=0.65, sigma=0.01)

curve(dgamma(x, prm[1], prm[2]), 0, 1) # prior for shape parameter
abline(v=0.65, lty=2, col='red')
}
\seealso{
Other random number generators: \code{\link{sim.gamma}}
}
\author{
John Giles
}
\concept{random number generators}
