% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.holidays}
\alias{get.holidays}
\title{Add national holidays and school breaks to long form data set}
\usage{
get.holidays(d, hol, sch)
}
\arguments{
\item{d}{expects output from \code{parse.long.form} (trip_durations_longform.csv)}

\item{hol}{expects data object \code{national.holidays} (national_holidays_clean.csv)}

\item{sch}{expects data object \code{school.terms} (school_terms_clean.csv)}
}
\value{
dataframe
}
\description{
This function takes output from the \code{parse.long.form} function and appends two columns with a binary response, 
where a 1 indicates that the date for that trip observation occured on a national holiday or a school break
}
\examples{
# Add indicator for trips that occur during national holidays or school breaks
tmp <- get.holidays(
     d=read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform.csv"),
     hol=read.csv('BeyondCommuting2/data/national_holidays_clean.csv', stringsAsFactors = FALSE),
     sch=read.csv('BeyondCommuting2/data/school_terms_clean.csv', stringsAsFactors = FALSE)
)
head(tmp)
}
\seealso{
Other data synthesis: \code{\link{calc.prop.tot.trips}},
  \code{\link{calc.route.type}},
  \code{\link{calc.samp.size}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.district.pop}},
  \code{\link{get.duration.counts}},
  \code{\link{get.subsamp}}, \code{\link{get.xy.counts}},
  \code{\link{mob.data.array}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
