% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.legend}
\alias{get.legend}
\title{Get legend from ggplot object}
\usage{
get.legend(x)
}
\arguments{
\item{x}{a ggplot object with a legend}
}
\value{
ggplot object
}
\description{
This function returns the legend of a ggplot object.
}
\seealso{
Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.admin.level}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.coords}()},
\code{\link{get.unique.names}()},
\code{\link{load.obj}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
