% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.param.vals}
\alias{get.param.vals}
\title{Get posterior parameter estimates from a model object}
\usage{
get.param.vals(n.districts, name = "lambda_1", stats, type = "matrix",
  n.cores = 2)
}
\arguments{
\item{n.districts}{Number of districts in model}

\item{name}{Name of the parameter (default = \code{lambda_1})}

\item{stats}{Expects statistics output of \code{\link{coda::summary}} function. If model out put is a \code{runjags} 
object, this can be given by: \code{summary(as.mcmc.list(out))$statistics}}

\item{type}{Return a matrix or dataframe (default = 'matrix')}

\item{n.cores}{Number of cores to use in parallel computation}
}
\value{
A list containing two matrices named \code{mean} and \code{sd}
}
\description{
A function to build a matrix or dataframe of the mean and standard deviation of the posterior distribution 
\strong{for each i to j route}.
}
\seealso{
Other model processing: \code{\link{jags.data.array.pop.level}},
  \code{\link{jags.data.array.route.level}},
  \code{\link{jags.data.array}}
}
\author{
John Giles
}
\concept{model processing}
