% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.pop.vec}
\alias{get.pop.vec}
\title{Get unique origin and destination names}
\usage{
get.pop.vec(data, orig = TRUE, dest = TRUE)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel.survey}} or derivative thereof}

\item{orig}{logical indicating whether to include origin location in unique names (default= TRUE)}

\item{dest}{logical indicating whether to include destination location in unique names (default= TRUE)}
}
\value{
two column dataframe containing unique names
}
\description{
This function returns coordinates for all unique locations in supplied data.
}
\seealso{
Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.admin.level}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.unique.coords}()},
\code{\link{get.unique.names}()},
\code{\link{load.obj}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
