% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.sparse.mob.matrix}
\alias{get.sparse.mob.matrix}
\title{Get sparse mobility matrix}
\usage{
get.sparse.mob.matrix(orig, dest, value, missing.obs = "NA")
}
\arguments{
\item{orig}{vector of origin names}

\item{dest}{vector of destination names}

\item{value}{vector of observed values for each origin/destination pair}

\item{missing.obs}{filler value for missing observations (default=NA)}
}
\value{
named numeric matrix
}
\description{
Takes X and Y coordinates of two locations and returns cross distance for all entries.
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.stay.data}()},
\code{\link{get.subsamp}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
