% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.stay.data}
\alias{get.stay.data}
\title{Make data set of travel and stays}
\usage{
get.stay.data(data, data.pred = NULL, agg.adm = NULL, name.class = "character")
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel.survey}} or derivative thereof}

\item{data.pred}{generalized data frame containing the admin units at which to predict probability of travel}

\item{agg.adm}{optional argument (logical) giving an arbitarary admin unit over which to aggregate travel/stay counts}

\item{name.class}{character indicating whether unique names should be either a unique character string (\code{name.class = "character"}) or a unique integer code (\code{name.class = "numeric"})}
}
\value{
dataframe with same columns as \code{\link{travel.survey}} data with columns for counts of travel/stay/total
}
\description{
This function build a data set containing the number of individuals that remain in their home location (stay) or travel to another location
during the time span of the survey. The output is designed to provide data for the \code{\link{fit.prob.travel}} function. The admin unit used to aggregate the travel/stay
counts depend on the arguemtns supplied:
\enumerate{
\item When \code{data.pred} is supplied and \code{agg.adm = NULL}, the lowest admin unit of \code{data.pred} is used
\item When both \code{data.pred} and \code{agg.adm} are \code{NULL}, the lowest admin unit of \code{data} is used
}
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.sparse.mob.matrix}()},
\code{\link{get.subsamp}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
