% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.subsamp}
\alias{get.subsamp}
\title{Find the subset of districts which have a minumim number of samples}
\usage{
get.subsamp(x, min.locations, min.samp)
}
\arguments{
\item{x}{a three dimensional array produced by the \code{mob.data.array} function}

\item{min.locations}{minimum number of locations (rows and columns) to keep}

\item{min.samp}{minimum sample size}
}
\value{
a three dimensional array containing the subset of \code{x}
}
\description{
This function takes the output from a the \code{mob.data.array} function and finds the 
largest subset of locations (districts) that have a minumum number of observations for all \eqn{ij} routes. The subset
is found by sequentially removing the location with the largest number of routes below the defined 
threshold (\code{min.samp}) until all locations contain at least \code{min.samp} number of observations for each route.
}
\examples{
load('./data/duration_data_arrays_3day_full.Rdata') # y.month, y.route, y.pop 

subsamp <- get.subsamp(y.route, min.locations=30, min.samp=10)

district.subset <- dimnames(subsamp)$origin
trip.map(as.numeric(district.subset))
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.sparse.mob.matrix}()},
\code{\link{get.xy.counts}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
