% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.unique.coords}
\alias{get.unique.coords}
\title{Get unique coordinates}
\usage{
get.unique.coords(data, orig = TRUE, dest = TRUE)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel.survey}} or derivative thereof}

\item{orig}{logical indicating whether to include unique coordinates from origin locations (default=TRUE)}

\item{dest}{logical indicating whether to include unique coordinates from destination locations (default=TRUE)}
}
\value{
three column dataframe containing unique coordinates and location names
}
\description{
This function returns the coordinates for all unique origins and/or destinations in supplied data.
}
\seealso{
Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.admin.level}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.names}()},
\code{\link{load.obj}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
