% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{get.unique.names}
\alias{get.unique.names}
\title{Get unique origin and destination names}
\usage{
get.unique.names(
  data,
  orig = TRUE,
  dest = TRUE,
  adm.start = NULL,
  adm.stop = NULL,
  name.class = "character"
)
}
\arguments{
\item{data}{generalized data frame described in \code{\link{travel.survey}}}

\item{orig}{logical indicating whether to include unique coordinates from origin locations (default=TRUE)}

\item{dest}{logical indicating whether to include unique coordinates from destination locations (default=TRUE)}

\item{adm.start}{highest administrative level to include in unique names (default = NULL, which uses highest observed in the data)}

\item{adm.stop}{lowest administrative level to include in unique names (default = NULL, which uses lowest observed in the data)}

\item{name.class}{character indicating whether unique names should be either a unique character string (\code{name.class = "character"}) or a unique integer code (\code{name.class = "numeric"})}
}
\value{
two column dataframe containing unique names
}
\description{
This function builds unique character string identifiers for all origins and destinations.
}
\seealso{
Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.admin.level}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.coords}()},
\code{\link{load.obj}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
