% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.xy.counts}
\alias{get.xy.counts}
\title{Get counts of any two variables (e.g. distance vs duration)}
\usage{
get.xy.counts(
  a,
  d,
  x.var,
  y.var = "duration",
  x.int,
  y.int,
  type = "matrix",
  distID = districtIDs,
  n.cores = NULL
)
}
\arguments{
\item{a}{origin district (can take integer ID or character name)}

\item{d}{longform mobility data, expects \code{d.42} data object (trip_durations_longform_metadata_42.csv) or similar}

\item{x.var}{name of x variable}

\item{y.var}{name of y variable (default = 'duration')}

\item{x.int}{interval for aggregating x variable}

\item{y.int}{interval for aggregating y variable}

\item{type}{return a 'matrix' or a 'dataframe' with counts in longform (default = 'matrix')}

\item{distID}{district IDs and names, expects \code{districtIDs} data object (NamNames.csv), default = districtIDs}

\item{n.cores}{number cores to use when parallel computing (default = NULL, which uses half available cores)}
}
\value{
matrix or dataframe
}
\description{
The \code{get.xy.counts} function counts the number of trips made conditioned on two variables (\code{x.var} and \code{y.var}) 
in the data set. This is set up to explore how trip duration counts vary with other variables in the data set 
(e.g. distance, district population size). Can be restricted to an origin district \code{a} or all origins \code{a} = NULL.
}
\examples{
data("d.42") # same as ./BeyondCommuting2/trip_lengths/trip_durations_longform_metadata_42.csv
data("districtIDs") # same as NamNames.csv

orig <- 42 # "Windhoek East"

# trip counts over distance and trip duration for an origin
m <- get.xy.counts(a=orig,               # integer ID of origin district
                   d=d.42,                  # expects "trip_durations_longform2.csv" or similar
                   x.var='distance',
                   y.var='duration',
                   x.int=50,             # distance interval for aggregating distance in km
                   y.int=50,             # duration interval for aggregating duration in days
                   type='matrix',
                   distID=districtIDs,
                   n.cores=4)
see(m)

plot3D::hist3D(x=as.numeric(rownames(m)), y=as.numeric(colnames(m)), z=log(m+1), 
               theta=120, phi=30, scale=F, expand=50,
               xlab='Distance',
               ylab='Duration',
               zlab='Log trip count',
               ticktype="detailed",
               space=0.05, lighting=T, light="diffuse")
}
\seealso{
Other data synthesis: 
\code{\link{calc.prop.tot.trips}()},
\code{\link{calc.route.type}()},
\code{\link{calc.samp.size}()},
\code{\link{get.crossdist}()},
\code{\link{get.distance.class}()},
\code{\link{get.distance.counts}()},
\code{\link{get.distance.matrix}()},
\code{\link{get.district.names.xy}()},
\code{\link{get.district.pop}()},
\code{\link{get.duration.counts}()},
\code{\link{get.holidays}()},
\code{\link{get.subsamp}()},
\code{\link{mob.data.array}()},
\code{\link{parse.longform}()}
}
\author{
John Giles
}
\concept{data synthesis}
