% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{jags.data.array}
\alias{jags.data.array}
\title{Build JAGS data array}
\usage{
jags.data.array(d, time = "month", variable = "duration",
  agg.int = 1)
}
\arguments{
\item{d}{Longform data with metadata attached}

\item{time}{The temporal interal used to construct the array (default = \code{'month'})}

\item{variable}{Character string giving the response variable; expects either \code{'distance'}, \code{'duration'}, or \code{'movement'}}

\item{agg.int}{The interval by which to aggregate the selected variable (default = 1). When \code{variable = 'duration'} this is the length 
of generation time (in days), when \code{variable = 'distance'} this is the distance interval in km. Ignored when \code{variable = 'movement'}.}
}
\value{
A 3-dimensional array when \code{variable = 'distance'} or \code{'movement'}, and a 4-dimensional array when \code{variable = 'duration'}
}
\description{
This function that builds data arrays at the time-level (e.g. month) for use in JAGS. The data arrays contain counts of either 
the distance or duration of trips, or the number of individuals moving from origin i to destination j.
}
\seealso{
Other model processing: \code{\link{get.param.vals}},
  \code{\link{jags.data.array.pop.level}},
  \code{\link{jags.data.array.route.level}}
}
\author{
John Giles
}
\concept{model processing}
