% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{jags.data.array.route.level}
\alias{jags.data.array.route.level}
\title{Reduce a jags.data.array object to route-level}
\usage{
jags.data.array.route.level(x, variable)
}
\arguments{
\item{x}{Month-level Output from the \code{\link{jags.data.array}} function}

\item{variable}{The variable in the data object (expects either \code{'distance'} or \code{'duration'})}
}
\value{
A 2-dimensional array when \code{variable = 'distance'}, and a 3-dimensional array when \code{variable = 'duration'}
}
\description{
This function reduces the output from the \code{\link{jags.data.array}} function---which contains data for the
month-level---to a route-level data array. The route-level array contains the mean value across all
months.
}
\seealso{
Other model processing: \code{\link{jags.data.array}}
}
\author{
John Giles
}
\concept{model processing}
