% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{jags.data.array2}
\alias{jags.data.array2}
\title{Build JAGS data array}
\usage{
jags.data.array2(d, time = "month", variable = "distance")
}
\arguments{
\item{a}{origin district (can take integer ID or character name)}
}
\value{
matrix or dataframe
}
\description{
This function that builds data arrays at the time-level (e.g. month) for use in JAGS. The data arrays contain counts of either 
the distance or duration of trips. When 'distance' is selected as the variable, the function returns 
a three-dimensional array and when 'duration' is selected, the function returns a four-dimensional array.
The difference in array dimensions occurs because the destination dimension must be collapsed in order to
perform counts of trip distance, leaving only the origin, time, and distance counts.
}
\author{
John Giles
}
\concept{model processing}
