% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{load.obj}
\alias{load.obj}
\title{Load specific objects in a .Rdata file}
\usage{
load.obj(x = 1, file)
}
\arguments{
\item{x}{the index corresponding to the desired object (default = 1)}

\item{file}{filepath to .Rdata object}
}
\value{
an R object
}
\description{
This function takes a .Rdata file and loads the indexed item(s) into the current environment as an object.
}
\examples{

sim <- load.obj("./output/TSIR_sim_influenza_high.Rdata")

y.route <- load.obj(3, './data/duration_data_arrays_1day_full.Rdata')  # y.route only
}
\seealso{
Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.admin.level}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.coords}()},
\code{\link{get.unique.names}()},
\code{\link{rjags.combine}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
