% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{mob.data.array.pop.level}
\alias{mob.data.array.pop.level}
\title{Reduce a mob.data.array object to population-level}
\usage{
mob.data.array.pop.level(x, variable)
}
\arguments{
\item{x}{Month-level output from the \code{\link{mob.data.array}} function}

\item{variable}{The variable in the data object (expects either \code{'distance'} or \code{'duration'})}
}
\value{
vector
}
\description{
This function reduces the output from the \code{\link{mob.data.array}} function---which contains data for the
month-level---to a vector giving the population mean. The population-level vector contains the observed mean 
value of the response variable across all routes and months.
}
\seealso{
Other model processing: \code{\link{calc.decay.stats}},
  \code{\link{get.param.vals}},
  \code{\link{mob.data.array.route.level}}
}
\author{
John Giles
}
\concept{model processing}
