% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{mob.data.array}
\alias{mob.data.array}
\title{Build mobility data array}
\usage{
mob.data.array(d, time = "month", variable = "duration", agg.int = 1,
  verbose = TRUE)
}
\arguments{
\item{d}{Longform data with metadata attached}

\item{time}{The temporal interal used to construct the array (default = \code{'month'})}

\item{variable}{Character string giving the response variable; expects either \code{'distance'}, \code{'duration'}, \code{'movement'}, or \code{'leave'}}

\item{agg.int}{an integer giving the interval by which to aggregate the selected variable (default = 1). When \code{variable = 'duration'} this is the length 
of generation time (in days), when \code{variable = 'distance'} this is the distance interval in km. Ignored when \code{variable = 'movement'}.}

\item{verbose}{print function updates (default = TRUE)}

\item{level}{The level of the data. If NULL, the level will be the same as the time variable. If set to \code{route} the mean across the time level will be taken for each route.}
}
\value{
A matrix when \code{variable = 'leave'}, 3-dimensional array when \code{variable = 'distance'} or \code{'movement'}, and a 4-dimensional array when \code{variable = 'duration'}
}
\description{
This function that builds data arrays at the time-level (e.g. month) for use in mob. The data arrays contain counts of either 
the distance or duration of trips, or the number of individuals moving from origin i to destination j.
}
\seealso{
Other data synthesis: \code{\link{calc.prop.route.type}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.district.pop}},
  \code{\link{get.duration.counts}},
  \code{\link{get.holidays}}, \code{\link{get.xy.counts}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
