% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{mob.data.array.route.level}
\alias{mob.data.array.route.level}
\title{Reduce a mob.data.array object to route-level}
\usage{
mob.data.array.route.level(x, variable)
}
\arguments{
\item{x}{Month-level Output from the \code{\link{mob.data.array}} function}

\item{variable}{The variable in the data object (expects either \code{'distance'}, \code{'duration'}, or \code{'movement'})}
}
\value{
A 2-dimensional array when \code{variable = 'distance'} or \code{'movement'}, and a 3-dimensional array when \code{variable = 'duration'}
}
\description{
This function reduces the output from the \code{\link{mob.data.array}} function---which contains data for the
month-level---to a route-level data array. The route-level array contains the mean value across all
months.
}
\seealso{
Other model processing: \code{\link{calc.decay.stats}},
  \code{\link{get.param.vals}},
  \code{\link{mob.data.array.pop.level}}
}
\author{
John Giles
}
\concept{model processing}
