% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{parse.longform}
\alias{parse.longform}
\title{Parse raw trip duration data}
\usage{
parse.longform(d, n.cores)
}
\arguments{
\item{d}{raw text file containing trip duration data}

\item{n.cores}{number cores to use when parallel computing}
}
\value{
a dataframe with the district IDs and names of the origin and destination districts, counts of each trip duration and detailed date information
}
\description{
This function takes the \code{d.raw} data object (trip_duration_counts.txt) or the \code{d.raw.42} data object (trip_durations_counts_sub_42.txt)
and parses the complex character string in the third column into a long form data set with date information. This function takes a while to run on the raw 
trip duration data, more cores are better.
}
\examples{
# process trip length data into long form data set with dates
t <- Sys.time()
tmp <- parse.longform(
     d=read.table('BeyondCommuting2/trip_lengths/trip_duration_counts.txt', header = FALSE, sep = "|"), 
     n.cores=6
)

head(tmp)

write.csv(tmp, file="BeyondCommuting2/trip_lengths/trip_durations_longform.csv", row.names=FALSE)
Sys.time() - t
}
\seealso{
Other data management: \code{\link{get.distance.class}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.district.pop}},
  \code{\link{get.holidays}}
}
\author{
John Giles
}
\concept{data management}
