% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_model_funcs.R
\name{rjags.combine}
\alias{rjags.combine}
\title{Combine two rjags objects}
\usage{
rjags.combine(a, b)
}
\arguments{
\item{x}{an rjags list}
}
\value{
an rjags list
}
\description{
This function combines two rjags objects.
}
\seealso{
Other model: 
\code{\link{fit.gravity}()},
\code{\link{fit.jags}()},
\code{\link{fit.prob.travel}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{summarize.hmob}()}

Other utility: 
\code{\link{check.distance.matrix}()},
\code{\link{get.admin.level}()},
\code{\link{get.legend}()},
\code{\link{get.names}()},
\code{\link{get.pop.vec}()},
\code{\link{get.unique.coords}()},
\code{\link{get.unique.names}()},
\code{\link{load.obj}()},
\code{\link{rjags.mcmc.list}()},
\code{\link{see}()},
\code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{model}
\concept{utility}
