% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{see}
\alias{see}
\title{Quick look at matrices}
\usage{
see(m, i = 6, j = 6)
}
\arguments{
\item{m}{matrix object}

\item{i}{number of rows to preview, default = 6}

\item{j}{number of columns to display, default = 6}
}
\value{
matrix
}
\description{
Displays the first 6 rows and columns of a matrix object
}
\examples{
tmp <- matrix(rnorm(100), nrow=10, ncol=10)

see(tmp)
see(tmp, 2, 3)


}
\seealso{
Other utility: \code{\link{get.legend}},
  \code{\link{get.names}}, \code{\link{load.obj}},
  \code{\link{trip.map}}
}
\author{
John Giles
}
\concept{utility}
