% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{sim.combine.dual}
\alias{sim.combine.dual}
\title{Combine TSIR scenarios containing both gravity model types}
\usage{
sim.combine.dual(x, y)
}
\arguments{
\item{x}{list object of simulation output from first iteration(s)}

\item{y}{list object of simulation output from second iteration(s)}
}
\value{
a list
}
\description{
Helper function designed to facilitate parallel implementation with the foreach package.This function combines
iterations of a TSIR scenario that contain output for both the basic gravity model (B) and the new gravity model
with duration (R).
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.age.beta}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
