% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{sim.combine}
\alias{sim.combine}
\title{Combine two TSIR scenarios}
\usage{
sim.combine(x, y)
}
\arguments{
\item{x}{list object of simulation output from first scenario}

\item{y}{list object of simulation output from second scenario}
}
\value{
a list
}
\description{
Combine output from two TSIR scenarios into one list object.
}
\seealso{
Other simulation: 
\code{\link{calc.hpd}()},
\code{\link{calc.prop.inf}()},
\code{\link{calc.prop.remain}()},
\code{\link{calc.timing.magnitude}()},
\code{\link{calc.wait.time}()},
\code{\link{decay.func}()},
\code{\link{get.age.beta}()},
\code{\link{get.beta.params}()},
\code{\link{sim.TSIR.full}()},
\code{\link{sim.TSIR}()},
\code{\link{sim.combine.dual}()},
\code{\link{sim.gravity.duration}()},
\code{\link{sim.gravity}()},
\code{\link{sim.lambda}()},
\code{\link{sim.pi}()},
\code{\link{sim.rho}()},
\code{\link{sim.tau}()}
}
\author{
John Giles
}
\concept{simulation}
