% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{sim.connectivity}
\alias{sim.connectivity}
\title{Simulate connectivity matrix}
\usage{
sim.connectivity(mu)
}
\arguments{
\item{mu}{a three dimensional array giving the mean of the posterior distribution for \eqn{\pi_{ijt}}}
}
\value{
A three dimensional array with values between 0 and 1, where rows (all \eqn{j} destination districts) sum to 1
}
\description{
This function takes the mean \eqn{\mu_\pi_{ijt}} of the estimated posterior distribution of \eqn{\pi_{ijt}} (the probability of movement 
from district \eqn{i} to district \eqn{i} in time \eqn{t}) and returns one stochastic realization of the connectivity matrix. Each stochastic 
realization is produced by returning the vector \eqn{\boldsymbol\pi_{i\{j\}t}} from a Dirichlet distribution.
}
\examples{
load('./output/gravity_model_8dists.Rdata')          # Gravity model parameters
tmp <- get.param.vals(n.districts=8, name='pi', level='month', stats=s.grav)
pi.hat <- sim.connectivity(mu=tmp$mean)
}
\seealso{
Other simulation: \code{\link{calc.p}},
  \code{\link{decay.func}}, \code{\link{get.beta.params}}
}
\author{
John Giles
}
\concept{simulation}
