% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{sim.gamma}
\alias{sim.gamma}
\title{Calculate the probability of a Gamma distributed random variable}
\usage{
sim.gamma(x, shape, rate)
}
\arguments{
\item{mu}{the desired mean of the Gamma distribution}

\item{sigma}{the desired standard deviation of the Gamma distribution}
}
\value{
a named numeric vector giving the \code{shape} and \code{rate} parameters of the Gamma distribution
}
\description{
A function that calcualtes the probability of random variable \code{x} according to a Gamma probability distribution with provided
\code{shape} and \code{rate} parameters.
}
\examples{

plot(sim.gamma(1:20, shape=1, rate=0.5))

plot(sim.gamma(1:100, shape=3, rate=0.05))
}
\seealso{
Other random number generators: 
\code{\link{get.gamma.params}()}
}
\author{
John Giles
}
\concept{random number generators}
