% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{sim.lambda}
\alias{sim.lambda}
\title{Simulate decay rate (lambda)}
\usage{
sim.lambda(mu, sigma, level)
}
\arguments{
\item{mu}{mean of posterior distribution of decay rate lambda}

\item{sigma}{standard deviation of posterior distribution of decay rate lambda}

\item{level}{the level of the data at which the mean and standard deviation of lambda are given (e.g. destination-, route- or month-level)}
}
\value{
a numerical matrix when \code{level = 'route'}
}
\description{
A function to simulate one stochastic realization of the decay rate lambda given the mean and standard deviation of the posterior 
distribution estimated by model of decay rate or a gravity model.
}
\examples{

load('./output/gravity_model_route_8dists.Rdata')    # mod.grav  #  posterior estimates of gravity model parameters
tmp <- get.param.vals(n.districts=8, name='lambda', level='route', stats=mod.grav)

lambda.hat <- sim.lambda(mu=tmp$mean, sigma=tmp$sd, level='route')
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.pi}},
  \code{\link{sim.rho}}, \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
