% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{sim.rho}
\alias{sim.rho}
\title{Simulate proportion remaining for full generation (rho)}
\usage{
sim.rho(p, level)
}
\arguments{
\item{p}{an array produced by the \code{calc.p} function giving the proportion of travellers remaining after \eqn{n} generations}

\item{level}{the level of the data for which to generate the stochastic realization of rho (e.g. destination-, route- or month-level)}
}
\value{
a numerical matrix when \code{level = 'route'}
}
\description{
A function that takes output from the \code{calc.p} function and simulates one stochastic realization
of rho. The function does so by calculating the mean and variance of \code{p} at the level indicated by the \code{level}
argument, and then derives the shape and rate parameters for the beta distribution.
}
\examples{
d <- read.csv("./data/trip_durations_longform_metadata_13_34_42_85_10_14_16_99_urban_rural.csv", stringsAsFactors=F)
d <- mob.data.array(d=d, time='month', variable='duration', agg.int=1)
p <- calc.prop.remain(d=d, gen.t=3)
rho.hat <- sim.rho(p=p, level='route')
}
\seealso{
Other simulation: 
\code{\link{calc.hpd}()},
\code{\link{calc.prop.inf}()},
\code{\link{calc.prop.remain}()},
\code{\link{calc.timing.magnitude}()},
\code{\link{calc.wait.time}()},
\code{\link{decay.func}()},
\code{\link{get.age.beta}()},
\code{\link{get.beta.params}()},
\code{\link{sim.TSIR.full}()},
\code{\link{sim.TSIR}()},
\code{\link{sim.combine.dual}()},
\code{\link{sim.combine}()},
\code{\link{sim.gravity.duration}()},
\code{\link{sim.gravity}()},
\code{\link{sim.lambda}()},
\code{\link{sim.pi}()},
\code{\link{sim.tau}()}
}
\author{
John Giles
}
\concept{simulation}
