% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{sim.tau}
\alias{sim.tau}
\title{Simulate probability of leaving origin (tau)}
\usage{
sim.tau(p)
}
\arguments{
\item{p}{A matrix giving the observed proportion of individuals of leaving origin \eqn{i} for each day in the trip suration data}
}
\value{
a numerical vector with values between 0 and 1
}
\description{
A function that takes a matrix containing the empirical proportion of individuals that leave origin \eqn{i}
for each date in the trip duration data and simulates one stochastic realization
of tau (the overall probability of leaving origin \eqn{i}). The function does so by calculating the mean and variance of \code{p} 
for each origin \eqn{i}, and then derives the shape and rate parameters for the beta distribution.
}
\examples{
load('./data/leave_data.Rdata')
load('./data/stay_data.Rdata')

prop.leave <- leave/stay

tau.hat <- sim.tau(prop.leave)
head(tau.hat)
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}}, \code{\link{load.obj}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}}
}
\author{
John Giles
}
\concept{simulation}
