% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.hpd}
\alias{calc.hpd}
\title{Calculate HPD of aggregated waiting time distributions}
\usage{
calc.hpd(x, ci = 0.95)
}
\arguments{
\item{x}{aggregated waiting time distributions for each district (output from \code{\link{calc.wait.time}})}

\item{ci}{a scalar or vector giving the HPD intervals to be calculated (default = 0.95)}
}
\value{
a dataframe
}
\description{
This function calculates the highest posterior density (HPD) interval for the aggregated 
waiting time distributions returned by the \code{\link{calc.wait.time}} function. The function 
calculates the maximum of each aggregated probability distribution along with the HPD intervals given by \code{ci}. 
Code adapted from the \code{hpd} function in the \code{TeachingDemos} package.
}
\examples{
load("./output/simulations/13/influenza.Rdata")

x <- sim$B$wait.time   # basic gravity model
#x <- sim$R$wait.time   # gravity model with duration 

agg <- calc.wait.time(x)
hpd <- calc.hpd(agg, ci=c(0.5, 0.95))

par(mfrow=c(3,4))
for (i in 1:nrow(hpd)) {
     plot(agg[i,], type='l', main=paste("district ", hpd$district[i]), ylab='Waiting time density')
     abline(v=hpd$lo95[i], lty=2, col='red')
     abline(v=hpd$hi95[i], lty=2, col='red')
     abline(v=hpd$max[i], col='red')
}

}
\seealso{
Other simulation: \code{\link{calc.prop.inf}},
  \code{\link{calc.prop.remain}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.age.beta}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}},
  \code{\link{sim.gravity.duration}},
  \code{\link{sim.gravity}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
