% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.p}
\alias{calc.p}
\title{Proportion of individuals remaining for full epidemic generation}
\usage{
calc.p(d, gen.t, n.gen = 1)
}
\arguments{
\item{d}{4-dimensional data array produced by the \code{\link{jags.data.array}} function \emph{or} the \code{duration.array.month.level} data object in the \pkg{hmobdata} package}

\item{gen.t}{the time interval in days used to define the epidemic generation}

\item{n.gen}{a scalar indicating the $n^th$ generation for which to calculate proportion of individuals remaining for the full epidemic generation}
}
\value{
A 4-dimensional array with values between 0 and 1
}
\description{
This function calculates the proportion of individuals that remain a location for
all of epidemic generation $n$. The value is represented by the variable \eqn{p_{jt}^{g(n)}}, which
is defined as the effective probability if individuals that stay for a full epidemic generation 
when they travel to destination \eqn{j} at time \eqn{t}:
\deqn{
p_{jt}^{g(n)} = \Pr(\text{remaining for all of n^{th} epidemic generation} \mid \text{generation time})
}
}
\seealso{
Other simulation: \code{\link{get.beta.params}}
}
\author{
John Giles
}
\concept{simulation}
