% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.p}
\alias{calc.p}
\title{Proportion of individuals remaining for full epidemic generation}
\usage{
calc.p(d, gen.t)
}
\arguments{
\item{d}{a four-dimensional array containing counts of trip durations produced by the \code{\link{mob.data.array}} function. 
Note that \code{calc.p} assumes that the duration data array is NOT aggregated (e.g. \code{mob.data.array} argument \code{agg.int}=1)}

\item{gen.t}{the time interval in days used to define the epidemic generation}
}
\value{
A 4-dimensional array with values between 0 and 1
}
\description{
This function calculates the proportion of individuals that remain a location for
all of epidemic generation \eqn{n}. The value is represented by the variable \eqn{p_jt}, which
is defined as the effective probability if individuals that stay for a full epidemic generation 
when they travel to destination \eqn{j} at time \eqn{t}:\cr
\cr
\eqn{p_ij} = Pr(remaining for all of \eqn{n^th} epidemic generation | generation time \eqn{g})
}
\examples{
library(hmobdata)
data(duration.array.month.level)

# Proportion travellers remaining after 1 generation, where the generation time is 3 days (influenza)
p <- calc.p(d=duration.array.month.level, gen.t=3, n.gen=1)

# Get destination-level mean and variance
p.mean <- apply(p, 2, mean, na.rm=TRUE)
p.var <- apply(p, 2, var, na.rm=TRUE)
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}}, \code{\link{load.obj}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
