% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.p}
\alias{calc.p}
\title{Proportion of individuals remaining for full epidemic generation}
\usage{
calc.p(d, gen.t, n.gen = 1)
}
\arguments{
\item{d}{a four-dimensional array containing counts of trip durations produced by the \code{\link{jags.data.array}} function}

\item{gen.t}{the time interval in days used to define the epidemic generation}

\item{n.gen}{a scalar indicating the $n^th$ generation for which to calculate proportion of individuals remaining for the full epidemic generation}
}
\value{
A 4-dimensional array with values between 0 and 1
}
\description{
This function calculates the proportion of individuals that remain a location for
all of epidemic generation \eqn{n}. The value is represented by the variable \eqn{p_jt}, which
is defined as the effective probability if individuals that stay for a full epidemic generation 
when they travel to destination \eqn{j} at time \eqn{t}:\cr
\cr
\eqn{p_ij} = Pr(remaining for all of \eqn{n^th} epidemic generation | generation time \eqn{g})
}
\examples{
library(hmobdata)
data(duration.array.month.level)

# Proportion travellers remaining after 1 generation, where the generation time is 3 days (influenza)
p <- calc.p(d=duration.array.month.level, gen.t=3, n.gen=1)

# Get destination-level mean and variance
p.mean <- apply(p, 2, mean, na.rm=TRUE)
p.var <- apply(p, 2, var, na.rm=TRUE)
}
\seealso{
Other simulation: \code{\link{decay.func}},
  \code{\link{get.beta.params}},
  \code{\link{sim.connectivity}}
}
\author{
John Giles
}
\concept{simulation}
