% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{calc.prop.route.type}
\alias{calc.prop.route.type}
\title{Calculate observed proportion of each route type}
\usage{
calc.prop.route.type(m, hi, lo)
}
\arguments{
\item{m}{A 3-dimensional data array produced by the \code{\link{mob.data.array}} function containing total trip counts (dimensions are origin, destination, time)}

\item{hi}{A vector of numerical district IDs in the high population density group}

\item{lo}{A vector of numerical district IDs in the low population density group}
}
\value{
A dataframe with four rows with values between 0 and 1, rows sum to 1
}
\description{
This function calculates the observed proportion of each of the four route types for 
each temporal unit in a mobility data array. The route types are defined by the population density 
of the origin and destination districts:
\enumerate{
  \item High density to high density (HH)
  \item High density to low density (HL)
  \item Low density to high density (LH)
  \item Low density to low density (LL)
}
}
\examples{
# Subset of 5 high pop density and 7 low pop density
d <- read.csv("BeyondCommuting2/trip_lengths/trip_durations_longform_metadata_12_dists_hi_lo_dens.csv", stringsAsFactors=F)
pop <- read.csv("./data/population_data.csv")

cutoff <- 980
hi <- pop$ID_2[pop$density >= cutoff]
lo <- pop$ID_2[pop$density < cutoff]

m <- mob.data.array(orig=d$from,
                    dest=d$to,
                    time=d$day, # total trip counts for each unique day
                    count=d$count,
                    name='movement')
str(m)

p <- calc.prop.route.type(m, hi, lo)

boxplot.matrix(as.matrix(p[,2:5]), col='lightblue')

cutoff <- 100
hi <- pop$ID_2[pop$density >= cutoff]
lo <- pop$ID_2[pop$density < cutoff]

p <- calc.prop.route.type(m, hi, lo)

boxplot.matrix(as.matrix(p[,2:5]), col='lightblue')


}
\seealso{
Other data synthesis: \code{\link{calc.prop.tot.trips}},
  \code{\link{calc.samp.size}},
  \code{\link{get.distance.class}},
  \code{\link{get.distance.counts}},
  \code{\link{get.district.names.xy}},
  \code{\link{get.district.pop}},
  \code{\link{get.duration.counts}},
  \code{\link{get.holidays}}, \code{\link{get.subsamp}},
  \code{\link{get.xy.counts}},
  \code{\link{mob.data.array}},
  \code{\link{parse.longform}}
}
\author{
John Giles
}
\concept{data synthesis}
