% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{decay.func}
\alias{decay.func}
\title{Exponential decay function}
\usage{
decay.func(alpha, lambda, y)
}
\arguments{
\item{alpha}{intercept (baseline number of expected trips at \eqn{y=0})}

\item{lambda}{decay rate}

\item{y}{a scalar or vector giving the data value(s) (e.g. trip duration)}
}
\value{
A scalar or vector of \eqn{N(y)}
}
\description{
This function calculates exponential decay of a data value \code{y} (e.g. trip duration)
given the intercept \eqn{\alpha} and decay rate \eqn{\lambda} parameters using the functional form:
\eqn{N(y) = \alpha exp(-\lambda*y)}. If \eqn{y =} duration, the function calculates the number of trips counted for a given duration.
}
\seealso{
Other simulation: \code{\link{calc.p}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
