% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.beta.params}
\alias{get.beta.params}
\title{Get parameters for Beta distribution}
\usage{
get.beta.params(m, v)
}
\arguments{
\item{m}{the mean \egn{\mu} of a random variable between 0 and 1}

\item{v}{the variance \egn{\sigma^2} of a random variable between 0 and 1}
}
\value{
A list containing \code{alpha} and \code{beta}
}
\description{
This function to calculates the two shape parameters (\code{alpha} and \code{beta}) for the Beta distribution 
using the mean \eqn{\mu} and variance \eqn{\sigma^2} of a variable that between 0 and one 1.
}
\seealso{
Other simulation: \code{\link{calc.p}}
}
\author{
John Giles
}
\concept{simulation}
