% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.beta.params}
\alias{get.beta.params}
\title{Get parameters for Beta distribution}
\usage{
get.beta.params(m, v)
}
\arguments{
\item{m}{the mean \eqn{\mu} of a random variable between 0 and 1}

\item{v}{the variance \eqn{\sigma^2} of a random variable between 0 and 1}
}
\value{
A list containing shape parameters \code{a} and \code{b}
}
\description{
This function to calculates the two shape parameters (\eqn{a} and \eqn{b}) for the Beta distribution 
using the mean \eqn{\mu} and variance \eqn{\sigma^2} of a variable that between 0 and one 1.
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}}, \code{\link{calc.p}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{load.obj}}, \code{\link{sim.TSIR.full}},
  \code{\link{sim.TSIR}}, \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
