% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{get.legend}
\alias{get.legend}
\title{Get legend from ggplot object}
\usage{
get.legend(x)
}
\arguments{
\item{x}{a ggplot object with a legend}
}
\value{
ggplot object
}
\description{
This function returns the legend of a ggplot object.
}
\seealso{
Other utility: \code{\link{get.names}()},
  \code{\link{load.obj}()}, \code{\link{see}()},
  \code{\link{trip.map}()}
}
\author{
John Giles
}
\concept{utility}
