% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{load.obj}
\alias{load.obj}
\title{Load .Rdata file to an object}
\usage{
load.obj(x = 1, file)
}
\arguments{
\item{x}{the index corresponding to the desired object (default = 1)}

\item{file}{filepath to .Rdata object}
}
\value{
an R object
}
\description{
This function a .Rdata file and loads the first item into the current environment as an object.
}
\examples{

sim <- load.obj("./output/TSIR_sim_influenza_high.Rdata")
}
\seealso{
Other simulation: \code{\link{calc.hpd}},
  \code{\link{calc.prop.inf}}, \code{\link{calc.p}},
  \code{\link{calc.timing.magnitude}},
  \code{\link{calc.wait.time}}, \code{\link{decay.func}},
  \code{\link{get.beta.params}},
  \code{\link{sim.TSIR.full}}, \code{\link{sim.TSIR}},
  \code{\link{sim.combine.dual}},
  \code{\link{sim.combine}}, \code{\link{sim.lambda}},
  \code{\link{sim.pi}}, \code{\link{sim.rho}},
  \code{\link{sim.tau}}
}
\author{
John Giles
}
\concept{simulation}
