% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmob_funcs.R
\name{mob.data.array.pop.level}
\alias{mob.data.array.pop.level}
\title{Reduce a mob.data.array object to population-level}
\usage{
mob.data.array.pop.level(x, variable)
}
\arguments{
\item{x}{Array output from the \code{\link{mob.data.array}} function}

\item{variable}{The variable in the data object (expects either \code{'distance'} or \code{'duration'})}
}
\value{
vector
}
\description{
This function reduces data array output from the \code{\link{mob.data.array}} function to the population-level by taking the mean trip count all routes and temporal intervals.
}
\examples{
# Month-level movement matrix
m <- mob.data.array(orig=df$from,
                    dest=df$to,
                    time=df$month,
                    count=df$count,
                    name='movement')
str(m)

# Mean-monthly movement for each route
m2 <- mob.data.array.pop.level(m, 'movement')


# Route-level distance matrix
m <- mob.data.array(orig=df$from,
                    dest=df$to,
                    count=df$count,
                    variable=df$distance,
                    agg.int=10, # count trips per 10km intervals
                    name='distance')

# Mean trip counts at each 10km distance interval for each origin
m2 <- mob.data.array.route.level(m, 'distance')


# Month-level aggregated duration matrix 
m <- mob.data.array(orig=df$from,
                    dest=df$to,
                    time=df$month,
                    count=df$count,
                    variable=df$duration,
                    agg.int=3, # aggregated to a generation time of 3 days
                    name='duration')
str(m)

# Mean monthly trip duration count for each route
m2 <- mob.data.array.route.level(m, 'duration')
}
\seealso{
Other model processing: 
\code{\link{calc.decay.stats}()},
\code{\link{get.param.vals}()},
\code{\link{mob.data.array.route.level}()}
}
\author{
John Giles
}
\concept{model processing}
