% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmobfuncs.R
\name{trip.map}
\alias{trip.map}
\title{Plot map of origin and destination district(s)}
\usage{
trip.map(a, b = NULL, distID = districtIDs, admin = adm, tol = 0.1)
}
\arguments{
\item{a}{origin district (can take integer ID or character name)}

\item{b}{destination district (can take integer ID or character name), default = NULL}

\item{distID}{district IDs and names, expacts the \code{districtIDs} data object (NamNames.csv)}

\item{tol}{tolerance argument passed to \code{gSimplify}, default = 0.1, set to zero for original}

\item{adm}{administrative level 2 data, expects NAM_adm2.shp}
}
\value{
plotted map
}
\description{
A function that maps the origin and destination(s) given district IDs
}
\examples{
par(mfrow=c(1,2))
trip.map(28, 20) # Generate map with district names
title(get.names(28, 20))
trip.map(42, c(16, 20, 1)) # Get map with multiple destinations
title(get.names(42))
}
\seealso{
Other utility: \code{\link{get.legend}()},
  \code{\link{get.names}()}, \code{\link{load.obj}()},
  \code{\link{see}()}
}
\author{
John Giles
}
\concept{utility}
